refactor trace type
