+++ src/app/(backend)/webapi/plugin/gateway/route.ts
@@ -1,9 +1,10 @@
 import { ChatErrorType, ErrorType } from '@lobechat/types/fetch';
+import { TraceNameMap } from '@lobechat/types/trace';
 import { PluginRequestPayload } from '@lobehub/chat-plugin-sdk';
 import { createGatewayOnEdgeRuntime } from '@lobehub/chat-plugins-gateway';
 
 import { LOBE_CHAT_AUTH_HEADER, OAUTH_AUTHORIZED, enableNextAuth } from '@/const/auth';
-import { LOBE_CHAT_TRACE_ID, TraceNameMap } from '@/const/trace';
+import { LOBE_CHAT_TRACE_ID } from '@/const/trace';
 import { getAppConfig } from '@/envs/app';
 import { AgentRuntimeError } from '@/libs/model-runtime';
 import { TraceClient } from '@/libs/traces';
