+++ src/features/AgentSetting/store/action.ts
@@ -1,12 +1,12 @@
-import type { PartialDeep } from 'type-fest';
+import { TraceNameMap, TracePayload, TraceTopicType } from '@lobechat/types/trace';
 import { getSingletonAnalyticsOptional } from '@lobehub/analytics';
+import type { PartialDeep } from 'type-fest';
 import { StateCreator } from 'zustand/vanilla';
 
 import { chainPickEmoji } from '@/chains/pickEmoji';
 import { chainSummaryAgentName } from '@/chains/summaryAgentName';
 import { chainSummaryDescription } from '@/chains/summaryDescription';
 import { chainSummaryTags } from '@/chains/summaryTags';
-import { TraceNameMap, TracePayload, TraceTopicType } from '@/const/trace';
 import { chatService } from '@/services/chat';
 import { useUserStore } from '@/store/user';
 import { systemAgentSelectors } from '@/store/user/slices/settings/selectors';
