+++ src/libs/traces/event.ts
@@ -1,7 +1,7 @@
+import { TraceEventType } from '@lobechat/types/trace';
 import { diffChars } from 'diff';
 import { LangfuseTraceClient } from 'langfuse-core';
 
-import { TraceEventType } from '@/const/trace';
 import {
   TraceEventBasePayload,
   TraceEventCopyMessage,
